refactor: consolidate MCP tool callback hierarchy and improve error handling
- Introduce AbstractMcpToolMethodCallback as common base class to eliminate code duplication
- Rename error handling methods for clarity:
  - createErrorResult -> createAsyncErrorResult/createSyncErrorResult
  - validateRequest -> validateSyncRequest
- Improve error messages using findCauseUsingPlainJava(e) to extract and display root cause instead of wrapper exceptions
