+++ mcp-annotations/src/main/java/org/springaicommunity/mcp/method/tool/AsyncStatelessMcpToolMethodCallback.java
@@ -77,7 +77,7 @@ public final class AsyncStatelessMcpToolMethodCallback extends AbstractAsyncMcpT
 			}
 			catch (Exception e) {
 				if (this.toolCallExceptionClass.isInstance(e)) {
-					return this.createErrorResult(e);
+					return this.createAsyncErrorResult(e);
 				}
 				return Mono.error(e);
 			}
